feat: add request id to tracing span
