+++ abci/Cargo.toml
@@ -23,13 +23,14 @@ server = [
 crypto = ["dep:lhash"]
 tcp = ["server"]
 unix = ["server"]
-tracing-span = []
+tracing-span = ["dep:uuid"]
 
 [[example]]
 name = "echo_socket"
 required-features = ["server"]
 
 [dependencies]
+uuid = { version = "1.4.1", features = ["v4", "fast-rng"], optional = true }
 tenderdash-proto = { version = "0.13.0-dev.2", path = "../proto" }
 bytes = { version = "1.0" }
 prost = { version = "0.11" }
